Use BigIntegerField for HyperParameter values

+++ app/src/bittensor_panel/core/models.py
@@ -5,4 +5,4 @@ class HyperParameter(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     name = models.CharField(max_length=255, unique=True)
-    value = models.IntegerField()
+    value = models.BigIntegerField()
